requirements_tensorflow-intel-gpu.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
ipykernel==6.30.0
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==3.19.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==311
pyzmq==27.0.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.10.0
termcolor==3.1.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
